+++ src/inspect_ai/_view/www/src/@types/log.d.ts
@@ -263,7 +263,7 @@ export type Type9 =
   | "permission"
   | "file_not_found"
   | "is_a_directory"
-  | "output_limit"
+  | "limit"
   | "approval"
   | "unknown";
 export type Message1 = string;